feat: "funcionarios" table updated and reload bug fixed
